requirements.txt
azure-ai-projects
azure-ai-inference[prompts]
azure-identity
azure-search-documents
pandas
python-dotenv
opentelemetry-api
markitdown
openai
azure-monitor-opentelemetry
azure-ai-evaluation[remote]
markdown
Flask>=2.2.0
gunicorn>=20.0.0
azure-identity
opentelemetry-api
opentelemetry-sdk
marshmallow==3.19.0